+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../ressources/styles/globals.scss'
+import '../styles/globals.scss'
 import 'antd/dist/antd.min.css'
 import type { AppProps } from 'next/app'
